backend/requirements.txt
Django
djangorestframework
reportlab
Pillow
django-cors-headers
psycopg2-binary
gunicorn
djangorestframework-simplejwt